refactor (informationmodel-cat): Move InformationModelHarvestSource under model
